refactor(kaizen): extract typed output accessors to OutputExtractionMixin
base_agent.py had grown to 1068 lines against the 1015 architectural
invariant in tests/regression/test_loc_invariants.py, so that guard was RED
on main.

Raising the limit was the disposition to avoid: it converts a guard into a
ratchet that records whatever happened rather than what was intended. The
sibling guard in test_base_agent_slimming.py shows where that leads — it has
been raised three times (1000 -> 1010 -> 1015 -> 1070) and can no longer fail
before the real invariant does.

So this restores the invariant the way the guard was written to force: by
extraction. The four extract_* accessors are the most separable unit in the
file — pure coercion over a result dict, reading no BaseAgent state at all
(they are callable with self=None, which is how the parity check was run).
They move verbatim into a mixin, matching the composition MCPMixin and
A2AMixin already established for this class rather than introducing a second
idiom.

Public API is unchanged: agent.extract_list/dict/float/str still resolve
through the MRO, so no callsite moves and no shim is needed. 1068 -> 1004.

Fixes #2119
